+++ coderadar-server/coderadar-core/src/test/java/org/wickedsource/coderadar/module/rest/ModuleControllerTest.java
@@ -38,7 +38,7 @@ public class ModuleControllerTest extends ControllerTestTemplate {
             document(
                 "modules/create",
                 requestFields(
-                    fields.withPath("id").description("The id of the module."),
+                    fields.withPath("id").description("The id of the module.").type("Long"),
                     fields
                         .withPath("modulePath")
                         .description(
