+++ coderadar-server/coderadar-core/src/test/java/org/wickedsource/coderadar/project/rest/ProjectControllerTest.java
@@ -57,7 +57,7 @@ public class ProjectControllerTest extends ControllerTestTemplate {
     return document(
         "projects/create",
         requestFields(
-            fields.withPath("id").description("The id of the project."),
+            fields.withPath("id").description("The id of the project.").type("Long"),
             fields.withPath("name").description("The name of the project to be analyzed."),
             fields
                 .withPath("vcsUrl")
